refactor : #207 Region vo 위치 수정
